Move tests to payload.rs
